more OSX 32/64 bit warnings removal...

+++ freex3d/src/lib/scenegraph/Tess.c
@@ -124,7 +124,7 @@ void CALLBACK FW_tess_error(GLenum e) {
 	/* Prints out tesselation errors. Older versions of at least MESA would
 	 give errors, so for now at least, lets just ignore them.
 	*/
-	 printf("FW_tess_error %d: >%s<\n",e,GL_ERROR_MSG); 
+	 printf("FW_tess_error %d: >%s<\n",(int) e,GL_ERROR_MSG); 
 }
